langchain-adapter/pyproject.toml
[project]
name = "langchain-adapter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.7.1",
    "isort>=6.0.1",
    "langchain-mcp-adapters>=0.1.7",
    "langchain[openai]>=0.3.25",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
]
